backend/requirements.txt
flask==3.0.3
flask-sock==0.7.0
opencv-python==4.10.0.84
numpy==2.1.2
scipy==1.13.1
matplotlib==3.9.2
pyserial==3.5
